update auth with jwt

+++ src/controllers/user/removeUser.js
@@ -1,11 +1,16 @@
 import userModel from "../../models/userModel.js";
 const removeUser = async (req, res) => {
-    const id = parseInt(req.params.id)
-    const morte = await prisma.userModel.deletear(id)
+  const id = parseInt(req.params.id);
 
-    return res.json({
-      message: `O Usuario ${id} Morreu`,
-    });
-    morte
+  if (id !== req.userLogged.id) {
+    return res.status(401).json({ error: "voce não pode matar os outros :(" });
   }
-  export default removeUser
+
+  const morte = await prisma.userModel.deletear(id);
+
+  return res.json({
+    message: `O Usuario ${id} Morreu`,
+  });
+  morte;
+};
+export default removeUser;
